# override float_16::to/fromFloat() with per-CPU optimized versions 