--- a/test/unit/FloatingPoint16/CMakeLists.txt
+++ b/test/unit/FloatingPoint16/CMakeLists.txt
@@ -6,7 +6,7 @@ add_onnx_mlir_executable(TestFloatingPoint16
   NO_INSTALL
 
   LINK_LIBS PRIVATE
-  OMONNXOps
+  OMMlirUtilities
   benchmark
   )