--- a/test/unit/FloatingPoint16/TestFloatingPoint16.cpp
+++ b/test/unit/FloatingPoint16/TestFloatingPoint16.cpp
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-//============================-- FloatingPoint16.cpp ---======================//
+//=======================-- TestFloatingPoint16.cpp ---=======================//
 //
 // Tests FloatingPoint16.
 //