When the CPU is known to support native conversion between float and float_16, use them in class float_16 below to override the slow default APFloat-based conversions.

Implemented for ARM64 when `__ARM_FP16_FORMAT_IEEE` is defined and for x86-64 when `__F16C__` is defined.

Added benchmarks of conversion between float and float_16 in TestFloatingPoint16.cpp.

Measured the benchmarks on a MacBook pro with Apple M1 Pro cpu. Before the per-CPU optimization:
```
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
BM_F32_TO_FP16<float_16>     2655030 ns      2654890 ns          263
BM_F32_TO_FP16<bfloat_16>    2741322 ns      2741320 ns          259
BM_FP16_TO_F32<float_16>     1824425 ns      1824426 ns          383
BM_FP16_TO_F32<bfloat_16>    1817702 ns      1817688 ns          385
```
After:
```
BM_F32_TO_FP16<float_16>      142300 ns       142280 ns         4855
BM_F32_TO_FP16<bfloat_16>    2312391 ns      2312290 ns          300
BM_FP16_TO_F32<float_16>      224026 ns       224026 ns         3124
BM_FP16_TO_F32<bfloat_16>    1832315 ns      1832315 ns          381
```

On another MacBook Pro with Intel Core i7-7820HQ cpu I had to configure the build with `-DCMAKE_CXX_FLAGS=-march=native` to define `__F16C__`. Before the per-CPU optimization:
```
BM_F32_TO_FP16<float_16>     4313245 ns      4304093 ns          140
BM_F32_TO_FP16<bfloat_16>    4427924 ns      4423289 ns          159
BM_FP16_TO_F32<float_16>     3426035 ns      3422583 ns          206
BM_FP16_TO_F32<bfloat_16>    3532166 ns      3517064 ns          204
```
After:
```
BM_F32_TO_FP16<float_16>       99106 ns        99069 ns         6525
BM_F32_TO_FP16<bfloat_16>    4358961 ns      4351302 ns          162
BM_FP16_TO_F32<float_16>      162100 ns       161966 ns         4204
BM_FP16_TO_F32<bfloat_16>    3459624 ns      3453559 ns          202
```