@AlexandreEichenberger 2023-06-14
Can there be some CMAKE magic that set these flags automatically when building the compiler on a machine that supports it? Or under a uniform flag that is common to all architectures (and default to nothing for arch not supporting it)?

Or if too complicated/not appropriate, can there be documentation (either in the cmakefile or in the .md doc) on how to enable this?