@sorenlassen 2023-06-14
> Can there be some CMAKE magic that set these flags automatically when building the compiler on a machine that supports it? Or under a uniform flag that is common to all architectures (and default to nothing for arch not supporting it)?
>
> Or if too complicated/not appropriate, can there be documentation (either in the cmakefile or in the .md doc) on how to enable this?

since there's a tradeoff between compile speed and portability, it may be best to make users make an explicit choice to target the faster and less portable CPU features

I added an explanation of how to set the CXX_FLAGS in BuildOnLinuxOSX.md

I can't test on Windows and don't know the right MSVC flags, so I'll leave that to anyone who wants to put that to use on Windows, if that's ok